backend/tools_deployment/knee-oa-cloudrun/requirements.txt
flask
gunicorn
pillow
numpy
onnxruntime